+++ README.md
@@ -7,17 +7,5 @@
 
 Если пример перестал собираться, попробуйте File -> Invalidate Caches / Restart
 
-Switch between branches to find sample you need:
--------------
-```shell
-v4-analytics-crash-kotlin
-v4-location-kotlin (incomplete)
-v4-pushkit-kotlin
-v4-remoteconfig-kotlin
-v4-clouddb-kotlin
-v4-iap-kotlin
-v4-server-pushkit-java
-v4-auth-service
-```
 [1]: https://github.com/huaweicodelabs/
 [2]: https://developer.huawei.com/consumer/en/doc/development/HMS-Guides/account-introduction-v4
