+++ app/src/main/java/com/huawei/hms/example/authservice/AnonymousLogin.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.huawei.hms.auth
+package com.huawei.hms.example.authservice
 
 import android.os.Bundle
 import android.view.View
