+++ app/src/main/res/layout/activity_email_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".EmailActivity">
+    tools:context="com.huawei.hms.example.authservice.EmailActivity">
 
     <EditText
         android:id="@+id/editTextEmail"
