+++ app/src/main/res/layout/activity_main.xml
@@ -5,13 +5,13 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context="com.huawei.hms.example.authservice.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        tools:context=".MainActivity">
+        tools:context="com.huawei.hms.example.authservice.MainActivity">
 
         <ImageView
             android:layout_width="match_parent"
@@ -40,7 +40,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"
-            tools:context=".MainActivity">
+            tools:context="com.huawei.hms.example.authservice.MainActivity">
 
             <Button
                 android:id="@+id/btnPhoneLogin"
