+++ app/src/main/res/layout/activity_phone_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".PhoneActivity">
+    tools:context="com.huawei.hms.example.authservice.PhoneActivity">
 
     <EditText
         android:id="@+id/editTextPhone"
